feedback

+++ src/native/minipal/ospagesize.h
@@ -4,6 +4,8 @@
 #ifndef HAVE_MINIPAL_OSPAGESIZE_H
 #define HAVE_MINIPAL_OSPAGESIZE_H
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -17,7 +19,7 @@ extern "C" {
 // On other platforms the value is queried from the OS once and cached; the
 // definition lives in ospagesize.c so there is exactly one cache per process.
 #if defined(__wasm__)
-static inline int minipal_getpagesize(void)
+static inline size_t minipal_getpagesize(void)
 {
     // The OS page size used by CoreCLR on WASM (16KB).
     // WASM has no hardware pages; getpagesize() returns the 64KB memory.grow granularity,
@@ -25,13 +27,13 @@ static inline int minipal_getpagesize(void)
     return 16 * 1024;
 }
 #elif defined(HOST_WINDOWS) || defined(_WIN32)
-static inline int minipal_getpagesize(void)
+static inline size_t minipal_getpagesize(void)
 {
     // The page size on Windows is 4KB and is not going to change.
     return 4 * 1024;
 }
 #else
-int minipal_getpagesize(void);
+size_t minipal_getpagesize(void);
 #endif
 
 #ifdef __cplusplus
